Please Add License
Please add this license to your repository as my work requires one as part of
our open source repository. As your goal is to make this the most popular Github
repo of all time I have included the [Unlicense](http://unlicense.org/) as an
example as it's widely considered to be the most open and permissive license.

Thank you.